src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/style-sharing-with-content-and-host-expected.txt

CONSOLE ERROR: line 14: Uncaught TypeError: document.querySelector(...).createShadowRoot is not a function
